backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
google-generativeai>=0.8.0
python-dotenv==1.0.0
Pillow==10.1.0
reportlab==4.0.7
python-docx==1.1.0
python-multipart==0.0.9
sqlmodel==0.0.16
python-jose[cryptography]==3.3.0
passlib==1.7.4
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-mock==3.11.1
